Izmene: pom.xml: dodat resources direktorijum Administrator.java: jmbg je long posto integer ne moze da sacuva 13 cifara Dostavljac.java: jmbg je long posto integer ne moze da sacuva 13 cifara GlavniEkran.java: dodate metode za postavljanje dostavljaca i administratora Login.java: ucitavanje podataka iz fajla i provera kredencijala RadSaFajlom.java: citanje fajlova i parsiranje podataka IzvorPodataka.java: podaci iz fajlova se ucitavaju u ovaj objekat i cuvaju u memoriji dok program radi administrator.txt i dostavljac.txt: teks fajlovi sa podacima

+++ src/main/java/ftn/vu/model/Dostavljac.java
@@ -2,15 +2,15 @@ package ftn.vu.model;
 
 public class Dostavljac extends Korisnik {
 	
-	private int jmbg;
+	private long jmbg;
 	
 	private double plata;
 	
-	public int getJmbg() {
+	public long getJmbg() {
 		return jmbg;
 	}
 
-	public void setJmbg(int jmbg) {
+	public void setJmbg(long jmbg) {
 		this.jmbg = jmbg;
 	}
